1.13

+++ src/main/kotlin/creep/CreepEndTask.kt
@@ -19,13 +19,19 @@ fun Creep.endTask(mainContext: MainContext) {
 
 
     //Break in dang room
-    if ((this.memory.role in 120..127 || this.memory.role in 1120..1125) && task.type != TypeOfTask.GoToRescueFlag) {
+    if ((this.memory.role in 120..127 || this.memory.role in 1120..1125)
+            && task.type != TypeOfTask.GoToRescueFlag
+            && task.type != TypeOfTask.GoToMainRoomRescueFlag) {
         val mainRoom: MainRoom? = mainContext.mainRoomCollector.rooms[this.memory.mainRoom]
         if (mainRoom == null) mainContext.tasks.deleteTask(this.id)
         else {
             val slaveRoom: SlaveRoom? = mainRoom.slaveRooms[this.memory.slaveRoom]
             if (slaveRoom == null) mainContext.tasks.deleteTask(this.id)
             else {
+                if (slaveRoom.constant.roomHostile) {
+                    mainContext.tasks.deleteTask(this.id)
+                    return
+                }
                 var indKl: Int = -1
                 if (this.memory.role in arrayOf(120, 121, 1120, 1121)) indKl = 0
                 if (this.memory.role in arrayOf(122, 123, 1122, 1123)) indKl = 1
